moved find_cluster to utils. follow_cluster now 1 task
