+++ pokemongo_bot/cell_workers/follow_cluster.py
@@ -1,7 +1,7 @@
 from pokemongo_bot import logger
 from pokemongo_bot.step_walker import StepWalker
-from pokemongo_bot.cell_workers import find_biggest_cluster
 from pokemongo_bot.cell_workers.utils import distance
+from pokemongo_bot.cell_workers.utils import find_biggest_cluster
 
 
 class FollowCluster(object):
@@ -10,19 +10,21 @@ class FollowCluster(object):
         self.is_at_destination = False
         self.announced = False
         self.dest = None
+        self.config = config
+        self._process_config()
+
+    def _process_config(self):
+        self.radius = self.config.get("radius", 50)
 
     def work(self):
-        worker = [x for x in self.bot.workers if type(x) == find_biggest_cluster.FindBiggestCluster][0]
-        self.radius = worker.radius
-
-        if worker.dest is not None:
-            self.dest = worker.dest
+        forts = self.bot.get_forts()
+        self.dest = find_biggest_cluster(self.radius, forts)
 
         if self.dest is not None:
 
             lat = self.dest['latitude']
             lng = self.dest['longitude']
-            cnt = self.dest['num_forts']
+            cnt = self.dest['num_points']
 
             if not self.is_at_destination:
                 log_str = 'Move to destiny. ' + str(cnt) + ' pokestops will in range of ' \
