+++ pokemongo_bot/cell_workers/utils.py
@@ -4,6 +4,10 @@ import struct
 from math import asin, atan, cos, exp, log, pi, sin, sqrt, tan
 
 from colorama import init
+from networkx.algorithms.clique import find_cliques
+
+import networkx as nx
+import numpy as np
 
 init()
 
@@ -207,3 +211,23 @@ def deg2rad(deg):
 
 def rad2deg(rad):
     return rad * 180.0 / pi
+
+
+def find_biggest_cluster(radius, points):
+    graph = nx.Graph()
+    for point in points:
+            f = point['latitude'], point['longitude']
+            graph.add_node(f)
+            for node in graph.nodes():
+                if node != f and distance(f[0], f[1], node[0], node[1]) <= radius*2:
+                    graph.add_edge(f, node)
+    cliques = list(find_cliques(graph))
+    if len(cliques) > 0:
+        max_clique = max(list(find_cliques(graph)), key=len)
+        merc_clique = [coord2merc(x[0], x[1]) for x in max_clique]
+        clique_x, clique_y = zip(*merc_clique)
+        best_point = np.mean(clique_x), np.mean(clique_y)
+        best_coord = merc2coord(best_point)
+        return {'latitude': best_coord[0], 'longitude': best_coord[1], 'num_points': len(max_clique)}
+    else:
+        return None
